gaia init automatic key processing
