+++ cmd/gaia/app/app.go
@@ -1,20 +1,14 @@
 package app
 
 import (
-	"encoding/json"
-	"errors"
+	"os"
 
-	"github.com/spf13/pflag"
-	"github.com/spf13/viper"
 	abci "github.com/tendermint/abci/types"
-	crypto "github.com/tendermint/go-crypto"
-	tmtypes "github.com/tendermint/tendermint/types"
 	cmn "github.com/tendermint/tmlibs/common"
 	dbm "github.com/tendermint/tmlibs/db"
 	"github.com/tendermint/tmlibs/log"
 
 	bam "github.com/cosmos/cosmos-sdk/baseapp"
-	"github.com/cosmos/cosmos-sdk/server"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/wire"
 	"github.com/cosmos/cosmos-sdk/x/auth"
@@ -25,6 +19,12 @@ import (
 
 const (
 	appName = "GaiaApp"
+)
+
+// default home directories for expected binaries
+var (
+	DefaultCLIHome  = os.ExpandEnv("$HOME/.gaiacli")
+	DefaultNodeHome = os.ExpandEnv("$HOME/.gaiad")
 )
 
 // Extended ABCI application
@@ -152,155 +152,4 @@ func (app *GaiaApp) initChainer(ctx sdk.Context, req abci.RequestInitChain) abci
 	stake.InitGenesis(ctx, app.stakeKeeper, genesisState.StakeData)
 
 	return abci.ResponseInitChain{}
-}
-
-//________________________________________________________________________________________
-
-// State to Unmarshal
-type GenesisState struct {
-	Accounts  []GenesisAccount   `json:"accounts"`
-	StakeData stake.GenesisState `json:"stake"`
-}
-
-// GenesisAccount doesn't need pubkey or sequence
-type GenesisAccount struct {
-	Address sdk.Address `json:"address"`
-	Coins   sdk.Coins   `json:"coins"`
-}
-
-func NewGenesisAccount(acc *auth.BaseAccount) GenesisAccount {
-	return GenesisAccount{
-		Address: acc.Address,
-		Coins:   acc.Coins,
-	}
-}
-
-// convert GenesisAccount to auth.BaseAccount
-func (ga *GenesisAccount) ToAccount() (acc *auth.BaseAccount) {
-	return &auth.BaseAccount{
-		Address: ga.Address,
-		Coins:   ga.Coins.Sort(),
-	}
-}
-
-var (
-	flagName = "name"
-	//flagOWK      = "overwrite-keys"
-)
-
-// get app init parameters for server init command
-func GaiaAppInit() server.AppInit {
-	fsAppGenState := pflag.NewFlagSet("", pflag.ContinueOnError)
-	//fsAppGenState.BoolP(flagOWK, "k", false, "overwrite the for the accounts created, if false and key exists init will fail")
-
-	fsAppGenTx := pflag.NewFlagSet("", pflag.ContinueOnError)
-	fsAppGenTx.String(flagName, "", "validator moniker, if left blank, do not add validator")
-
-	return server.AppInit{
-		FlagsAppGenState: fsAppGenState,
-		FlagsAppGenTx:    fsAppGenTx,
-		AppGenState:      GaiaAppGenState,
-		AppGenTx:         GaiaAppGenTx,
-	}
-}
-
-// simple genesis tx
-type GaiaGenTx struct {
-	Name    string        `json:"name"`
-	Address sdk.Address   `json:"address"`
-	PubKey  crypto.PubKey `json:"pub_key"`
-}
-
-// power given to validators in gaia init functions
-var FreePower = int64(100)
-
-// Create the core parameters for genesis initialization for gaia
-// note that the pubkey input is this machines pubkey
-func GaiaAppGenState(cdc *wire.Codec, appGenTxs []json.RawMessage) (appState json.RawMessage, err error) {
-
-	if len(appGenTxs) == 0 {
-		err = errors.New("must provide at least genesis transaction")
-		return
-	}
-
-	// start with the default staking genesis state
-	stakeData := stake.GetDefaultGenesisState()
-
-	// get genesis flag account information
-	genaccs := make([]GenesisAccount, len(appGenTxs))
-	for i, appGenTx := range appGenTxs {
-
-		var genTx GaiaGenTx
-		err = cdc.UnmarshalJSON(appGenTx, &genTx)
-		if err != nil {
-			return
-		}
-
-		// create the genesis account, give'm few fermions and a buncha token with there name
-		accAuth := auth.NewBaseAccountWithAddress(genTx.Address)
-		accAuth.Coins = sdk.Coins{
-			{genTx.Name + "Token", 1000},
-			{"fermion", 50},
-		}
-		acc := NewGenesisAccount(&accAuth)
-		genaccs[i] = acc
-
-		// add the validator
-		if len(genTx.Name) > 0 {
-			desc := stake.NewDescription(genTx.Name, "", "", "")
-			candidate := stake.NewCandidate(genTx.Address, genTx.PubKey, desc)
-			candidate.Assets = sdk.NewRat(FreePower)
-			stakeData.Candidates = append(stakeData.Candidates, candidate)
-
-			// pool logic
-			stakeData.Pool.TotalSupply += FreePower
-			stakeData.Pool.BondedPool += FreePower
-			stakeData.Pool.BondedShares = sdk.NewRat(stakeData.Pool.BondedPool)
-		}
-	}
-
-	// create the final app state
-	genesisState := GenesisState{
-		Accounts:  genaccs,
-		StakeData: stakeData,
-	}
-	appState, err = wire.MarshalJSONIndent(cdc, genesisState)
-	return
-}
-
-// Generate a gaia genesis transaction
-func GaiaAppGenTx(cdc *wire.Codec, pk crypto.PubKey) (
-	appGenTx, cliPrint json.RawMessage, validator tmtypes.GenesisValidator, err error) {
-
-	var addr sdk.Address
-	var secret string
-	addr, secret, err = server.GenerateCoinKey()
-	if err != nil {
-		return
-	}
-
-	var bz []byte
-	gaiaGenTx := GaiaGenTx{
-		Name:    viper.GetString(flagName),
-		Address: addr,
-		PubKey:  pk,
-	}
-	bz, err = wire.MarshalJSONIndent(cdc, gaiaGenTx)
-	if err != nil {
-		return
-	}
-	appGenTx = json.RawMessage(bz)
-
-	mm := map[string]string{"secret": secret}
-	bz, err = cdc.MarshalJSON(mm)
-	if err != nil {
-		return
-	}
-	cliPrint = json.RawMessage(bz)
-
-	validator = tmtypes.GenesisValidator{
-		PubKey: pk,
-		Power:  FreePower,
-	}
-	return
 }
